requirements.txt
streamlit>=1.24
langchain>=0.0.365
chromadb>=0.3.33
python-docx>=0.8.11
pypdf>=3.9.0
openai>=0.27.0
tqdm
python-multipart
sentence-transformers>=2.2.2
pdfplumber
python-dotenv